Update Excel-to-Word-index-converter.ps1
Moved initialization of MS Word Document and adding contents to MS Word Document into functions. This is to improve ability to add support for other document types in a later version.

Chose not to split these out into separate modules at this time. That may happen as the script grows.

Signed-off-by: James VanOeffelen <47988842+VanAwful@users.noreply.github.com>